requirements.txt
obd==0.7.1
pynmea2==1.19.0
pyserial==3.4
requests==2.31.0
websockets==11.0.3
